Shop owner and item tables joined using One to many mapping.

+++ src/main/java/com/shop/repositary/ShopOwnerRepo.java
@@ -6,7 +6,7 @@ import org.springframework.stereotype.Repository;
 import com.shop.model.ShopOwner;
 
 @Repository
-public interface ShopOwnerRepo extends JpaRepository<ShopOwner, Integer>{
+public interface ShopOwnerRepo extends JpaRepository<ShopOwner, Long>{
